requirements.txt
streamlit==1.23.1
openai==1.36.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
python-dotenv==1.0.1